build.rs: Unify the way generated sources are processed.
If we're not using pregenerated sources, then just immediately
generate the sources and then continue on using the same logic as we
use when we are using pregenerated sources. This makes things easier
to understand and maintain. This also allows us to simplify the build
logic for .S files on Aarch64 Windows in the future.