ChaCha20-Poly1305 internals: Clarify memory safety of `encrypt_within`.
Eliminate the "less safe" variant of `encrypt_within`. Move the check
for overlapping buffers into the inner safe wrapper around the assembly
function call, so that it is clear what we're giving the assembly
function.

The extra checks are only done for 32-bit ARM and 32-bit x86, which
are less of a performance priority now. And also the checks probably
don't affect performance anyway.